# openapi.yaml
openapi: 3.1.0
info:
  title: INAM Protocol Registry API
  version: "0.33.0"
  description: >
    The open reputation, verification, and economic-history layer for the agent
    economy. See SPEC.md in this repository for the full protocol specification
    — this document describes the REST surface only. Complements MCP, A2A,
    AgentPass, AITP, and Passport Alliance rather than replacing any of them.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.inamprotocol.org/v1
    description: Live Cloudflare Workers deployment (reference instance)
  - url: http://localhost:4021/v1
    description: Local Node reference server (npm run dev)
...
paths:
...
  /agents:
    post:
      operationId: registerAgent
      summary: Register the calling INAM ID
      tags: [agents]
      security: [{ inamSignature: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capabilities]
              properties:
                capabilities: { type: array, items: { type: string }, minItems: 1 }
                metadata: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /agents/search:
    get:
      operationId: searchAgents
      summary: Discover agents by capability, reputation, or supported external protocol
      tags: [agents]
      security: []
      parameters:
        - name: capability
          in: query
          schema: { type: string }
        - name: min_reputation
          in: query
          schema: { type: number }
        - name: supports
          in: query
          description: One of agentpass_id, aitp_id, passport_id, erc8004_id, a2a_endpoint
          schema: { type: string }
        - name: include_revoked
          in: query
          description: "Set to `true` to include revoked agents (SPEC.md §2.2); excluded by default."
          schema: { type: boolean }
        - name: include_demo
          in: query
          description: "Set to `true` to include agents registered with `metadata.demo: true` (SPEC.md §2, v0.33); excluded by default."
          schema: { type: boolean }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200":
          description: Matching agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents: { type: array, items: { $ref: "#/components/schemas/Agent" } }
                  hasMore: { type: boolean }
        "429": { $ref: "#/components/responses/Error" }
  /agents/{id}:
    get:
      operationId: getAgent
      summary: Fetch an agent's public profile
      tags: [agents]
      security: []
      parameters: [{ $ref: "#/components/parameters/AgentId" }]
      responses:
        "200":
          description: Agent profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{id}/protocols:
    get:
      operationId: getAgentProtocols
      summary: Fetch an agent's linked external identities
      tags: [agents]
      security: []
      parameters: [{ $ref: "#/components/parameters/AgentId" }]
      responses:
        "200":
          description: Linked identities
          content:
            application/json:
              schema:
                type: object
                properties:
                  linked: { $ref: "#/components/schemas/LinkedIdentities" }
                  linkedProof: { $ref: "#/components/schemas/LinkedIdentityProofs" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{id}/reputation:
    get:
      operationId: getReputation
      summary: Compute and return an agent's reputation
      tags: [agents]
      security: []
      parameters: [{ $ref: "#/components/parameters/AgentId" }]
      responses:
        "200":
          description: Reputation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reputation" }
        "404": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /agents/{id}/badge.svg:
    get:
      operationId: getReputationBadgeSvg
      summary: Embeddable shields.io-style SVG badge showing an agent's live trust score
      description: >
        Read-only, unsigned, public. Renders only a fixed "inam" label and a
        server-computed score/status — never agent-supplied text (e.g.
        metadata.name) — so it's safe to embed as `<img src=...>` from any
        origin (CORS is open on this route). An unregistered id:key still
        returns 200 with a distinct grey "unknown" badge rather than a 404,
        so the `<img>` never renders as broken; a brand-new agent with no
        receipt history yet gets its own neutral grey "new" badge instead of
        being visually implied to have a bad (red) score. Example:
        `![reputation](https://api.inamprotocol.org/v1/agents/<did>/badge.svg)`.
      tags: [agents]
      security: []
      parameters: [{ $ref: "#/components/parameters/AgentId" }]
      responses:
        "200":
          description: SVG badge image (always 200; unknown/new agents get a distinct badge, not an error)
          content:
            image/svg+xml:
              schema: { type: string }
        "429": { $ref: "#/components/responses/Error" }
  /agents/{id}/badge.json:
    get:
      operationId: getReputationBadgeJson
      summary: Same badge data as badge.svg, as JSON (for building a custom badge renderer)
      description: >
        Follows shields.io's "endpoint badge" schema (schemaVersion/label/message/color),
        so this URL can also be passed directly to
        `https://img.shields.io/endpoint?url=<this URL>`.
      tags: [agents]
      security: []
      parameters: [{ $ref: "#/components/parameters/AgentId" }]
      responses:
        "200":
          description: Badge data (always 200; see badge.svg for the not-found/new-agent handling)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BadgeData" }
        "429": { $ref: "#/components/responses/Error" }
  /agents/{id}/receipts:
    get:
      operationId: listAgentReceipts
      summary: List an agent's receipts (draft, finalized, disputed)
      tags: [agents]
      security: []
      parameters: [{ $ref: "#/components/parameters/AgentId" }]
      responses:
        "200":
          description: Receipts
          content:
            application/json:
              schema:
                type: object
                properties: { receipts: { type: array, items: { $ref: "#/components/schemas/Receipt" } } }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{id}/verifier-status:
    post:
      operationId: setVerifierStatus
      summary: Grant or revoke an agent's verifier status (operator only, SPEC.md §12.6)
      description: >-
        Signed by the registry's configured operator identity, never by the
        target agent. Any other caller gets 403 NOT_OPERATOR; a registry with
        no operator configured rejects every call.
      tags: [agents]
      security: [{ inamSignature: [] }]
      parameters:
        - { $ref: "#/components/parameters/AgentId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorized]
              properties: { authorized: { type: boolean } }
      responses:
        "200":
          description: Updated agent record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{id}/revoke:
    post:
      operationId: revokeAgent
      summary: One-way retire this INAM ID (SPEC.md §2.2)
      description: >-
        Signed, self only. Sets `revokedAt`/`revocationReason` on the record.
        After this the ID performs no further signed operations (403
        AGENT_REVOKED), drops out of search, and is flagged `revoked` in its
        reputation. One-way — no un-revoke.
      tags: [agents]
      security: [{ inamSignature: [] }]
      parameters:
        - { $ref: "#/components/parameters/AgentId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string, minLength: 1, maxLength: 500 } }
      responses:
        "200":
          description: Updated (revoked) agent record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /agents/{id}/link/challenge:
    post:
      operationId: requestLinkChallenge
      summary: Request a single-use proof-of-control challenge before linking a key-derived external identity (SPEC.md §2.1)
      tags: [agents]
      security: [{ inamSignature: [] }]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [protocol, externalPublicKey, keyType]
              properties:
                protocol: { type: string, enum: [agentpass_id, aitp_id, passport_id, erc8004_id] }
                externalPublicKey: { type: string, description: "Base64-encoded external public key (uncompressed SEC1 point for secp256k1)" }
                keyType: { type: string, enum: [ed25519, p256, secp256k1] }
      responses:
        "201":
          description: Challenge issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkChallenge" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /agents/{id}/link:
    post:
      operationId: linkIdentity
      summary: Claim an external identity — agentpass_id/aitp_id/passport_id/erc8004_id require a completed challenge (SPEC.md §2.1); a2a_endpoint does not, since it isn't a key-derived identity
      tags: [agents]
      security: [{ inamSignature: [] }]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [protocol, value]
              properties:
                protocol: { type: string, enum: [agentpass_id, aitp_id, passport_id, erc8004_id, a2a_endpoint] }
                value: { type: string, description: "For erc8004_id, MUST equal the Ethereum address derived from externalPublicKey (SPEC.md §2.1) — otherwise rejected as ERC8004_ID_MISMATCH" }
                challengeId: { type: string, description: "Required for agentpass_id/aitp_id/passport_id/erc8004_id — from a prior /link/challenge call" }
                proofSignature: { type: string, description: "Base64 signature over the challenge, from the external private key" }
      responses:
        "200":
          description: Updated agent profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    inamSignature:
      type: apiKey
      in: header
      name: inam-signature
      description: >
        Not a static API key — see SPEC.md §6. Every signed request also
        requires `inam-agent` (the caller's INAM ID) and `inam-timestamp`
        (unix ms, within a 5-minute skew window) headers, and the signature
        itself is computed over `${METHOD}\n${fullPath}\n${timestamp}\n${sha256hex(rawBody)}`.
  parameters:
    AgentId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: An INAM ID (did:key:z...)
      example: "did:key:z6Mkon3Necd6NkkyfoGoHxid2znGc59LU3K7mubaRcFbLfLX"
...
    Limit:
      name: limit
      in: query
      description: "Page size (SPEC.md §6, v0.28). Defaults to 50, clamped to 200; invalid values fall back to the default."
      schema: { type: integer, default: 50, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: >-
        Required on every mutating request. A repeated (caller, key) pair replays the cached
        response of the prior *successful* (2xx) operation; a non-2xx is not cached and a retry
        re-executes. A given request signature is bound to one key — replaying a captured
        signature with a different key is rejected 409 REPLAYED_REQUEST (SPEC.md §7).
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    LinkedIdentities:
      type: object
      properties:
        agentpass_id: { type: string }
        aitp_id: { type: string }
        passport_id: { type: string }
        erc8004_id: { type: string, description: "Ethereum address, 0x-prefixed lowercase hex (SPEC.md §2.1)" }
        a2a_endpoint: { type: string }
    LinkProof:
      type: object
      description: >-
        How one `linked` entry was verified (SPEC.md §2.1). `key_possession`:
        proved control of `externalPublicKey` at `verifiedAt` — NOT a claim the
        key is authoritative for the identity externally (cross-registry
        resolution is out of scope, §10). `unverified_claim`: INAM-signed
        assertion only (`a2a_endpoint`).
      properties:
        method: { type: string, enum: [key_possession, unverified_claim] }
        verifiedAt: { type: string, format: date-time }
        keyType: { type: string, enum: [ed25519, p256, secp256k1], description: "present iff method is key_possession" }
        externalPublicKey: { type: string, description: "base64; present iff method is key_possession" }
      required: [method, verifiedAt]
    LinkedIdentityProofs:
      type: object
      description: Per-protocol assurance metadata for `linked`, same keys.
      properties:
        agentpass_id: { $ref: "#/components/schemas/LinkProof" }
        aitp_id: { $ref: "#/components/schemas/LinkProof" }
        passport_id: { $ref: "#/components/schemas/LinkProof" }
        erc8004_id: { $ref: "#/components/schemas/LinkProof" }
        a2a_endpoint: { $ref: "#/components/schemas/LinkProof" }
    LinkChallenge:
      type: object
      properties:
        challengeId: { type: string }
        challenge: { type: string, description: "64 hex chars — 32 random bytes to sign with the external private key" }
        expiresAt: { type: string, format: date-time }
    Agent:
      type: object
      properties:
        id: { type: string }
        capabilities: { type: array, items: { type: string } }
        metadata: { type: object, additionalProperties: true }
        linked: { $ref: "#/components/schemas/LinkedIdentities" }
        linkedProof: { $ref: "#/components/schemas/LinkedIdentityProofs" }
        stakeUsd: { type: number }
        isAuthorizedVerifier: { type: boolean, description: "SPEC.md §12.3 — operator-granted; false at registration" }
        revokedAt: { type: string, format: date-time, description: "SPEC.md §2.2 — set when the agent retired this ID; absent for an active agent" }
        revocationReason: { type: string }
        createdAt: { type: string, format: date-time }
    Reputation:
      type: object
      properties:
        trustScore: { type: number, minimum: 0, maximum: 100 }
        evidenceLevel:
          type: string
          enum: [none, countersigned, independently_verified]
          description: "Strongest evidence behind this history (SPEC.md §5.3, v0.32). Read this before relying on trustScore: 'countersigned' means only the two parties vouched for the work."
        components:
          type: object
          properties:
            eigenWeight: { type: number, description: "Confidence — how much trust-weighted history backs this score" }
            finalizedReceipts: { type: integer, description: "Receipts both parties signed — not independently verified (see attestedReceipts)" }
            verifiedReceipts: { type: integer, deprecated: true, description: "Deprecated alias of finalizedReceipts" }
            rawReceipts: { type: integer }
            successRate: { type: number }
            volumeUsd: { type: number, description: "Self-reported settlement volume in USD only (the 'USD' bucket of volumeByCurrency); INAM verifies no payment and does no FX (SPEC.md §5.3, §10)" }
            volumeByCurrency:
              type: object
              additionalProperties: { type: number }
              description: "Settlement volume bucketed by normalized settlement.currency (upper-cased; untagged -> 'USD'), never converted or cross-summed"
            stakeUsd: { type: number }
            decayHalfLifeDays: { type: number }
            attestedReceipts: { type: integer, description: "Finalized receipts whose authorized Verifications net out to verified (SPEC.md §12.5)" }
            rejectedAttestations: { type: integer, description: "Finalized receipts whose authorized Verifications net out to rejected; each is scored as a failed outcome (SPEC.md §12.5, v0.32)" }
            nonPerformanceReports: { type: integer, description: "Jobs where this agent was the accepted worker and the poster later reported non-performance (SPEC.md §3.3)" }
            asProvider: { $ref: "#/components/schemas/ReputationRoleBreakdown" }
            asRequester: { $ref: "#/components/schemas/ReputationRoleBreakdown" }
        flags: { type: array, items: { type: string }, description: "e.g. in_dispute, concentrated_counterparty:<did>, nonperformance_reported, attestation_rejected, unanchored_counterparty_volume, revoked" }
    ReputationRoleBreakdown:
      type: object
      description: "Weighted receipt count / success rate / volume for one side of an agent's history (SPEC.md §5.3), split by role held on each finalized receipt"
      properties:
        receipts: { type: integer }
        successRate: { type: number }
        volumeUsd: { type: number }
        volumeByCurrency:
          type: object
          additionalProperties: { type: number }
    BadgeData:
      type: object
      description: shields.io "endpoint badge" schema, plus INAM-specific status/trustScore
      properties:
        schemaVersion: { type: integer, enum: [1] }
        label: { const: inam }
        message: { type: string, description: "Formatted trustScore, or \"new\" / \"unknown\"" }
        color: { type: string, description: "Hex color, no leading #" }
        status: { type: string, enum: [scored, new, not_found] }
        trustScore: { type: number, minimum: 0, maximum: 100, description: "Omitted when status is not_found" }
...
    Receipt:
      type: object
      properties:
        receiptVersion: { const: "1.0" }
        receiptId: { type: string }
        jobId: { type: string }
        agentA: { type: object, properties: { id: { type: string }, role: { const: requester } } }
        agentB: { type: object, properties: { id: { type: string }, role: { const: worker } } }
        task: { type: object }
        result: { type: object }
        settlement: { type: object }
        verification:
          type: object
          description: "The receipt's own self-declared claim of how the work was checked (SPEC.md §4.1) — not the same resource as Verification (§12, the /verifications endpoints below), which is a separate, independently-signed third-party attestation with its own distinct `method` enum."
          properties:
            method: { type: string, enum: [payer_confirmation, independent_validator, test_suite_pass], description: "Self-declared by agent_b, unenforced today except payer_confirmation. Not Verification.method (§12) below." }
            verifier: { type: string }
            outcome: { type: string, enum: [success, partial, failed] }
        dispute:
          type: object
          properties:
            status: { type: string, enum: [none, open, resolved] }
            reason: { type: string }
            windowClosesAt: { type: [string, "null"], format: date-time, description: "null while the receipt is a draft; set when it is countersigned (SPEC.md §4.3, v0.31)" }
            openedBy: { type: string, description: "INAM ID of the party that opened the dispute (SPEC.md §4.3); absent until disputed" }
            resolvedAt: { type: string, format: date-time, description: "set when the opener withdrew the dispute" }
            resolution: { type: string, description: "optional note the opener left on withdrawal" }
        signatures:
          type: object
          properties: { agentB: { type: string }, agentA: { type: string } }
        status: { type: string, enum: [draft, finalized, disputed] }